+++ src/modules/meta/base.cpp
@@ -139,6 +139,14 @@ namespace modules {
     m_formats.insert(make_pair(move(name), move(format)));
   }
 
+  void module_formatter::add_default_formats(string fallback, vector<string>&& tags, vector<string>&& whitelist) {
+
+    add(DEFAULT_FORMAT, fallback, forward<vector<string>>(tags), forward<vector<string>>(whitelist));
+
+    // Default FORMAT-HOVER shall be the same as FORMAT for backwards compatibility
+    add(DEFAULT_FORMAT_HOVER, get(DEFAULT_FORMAT)->value, forward<vector<string>>(tags), forward<vector<string>>(whitelist));
+  }
+
   void module_formatter::add(string name, string fallback, vector<string>&& tags, vector<string>&& whitelist) {
     string value = m_conf.get(m_modname, name, move(fallback));
     add_value(move(name), move(value), forward<vector<string>>(tags), forward<vector<string>>(whitelist));