+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `custom/script`: `hover-start` and `hover-end` actions ([`Polybar #1064`](https://github.com/polybar/polybar/issues/1064), [`Polybar #2868`](https://github.com/polybar/polybar/pull/2868)).
   - Added support for TAG_LABEL (`<label>`) in ipc module  ([`Polybar #2841`](https://github.com/polybar/polybar/pull/2841))  by [@madhavpcm](https://github.com/madhavpcm).
   - Added support for format-i for each hook-i defined in ipc module ([`Polybar #2775`](https://github.com/polybar/polybar/issues/2775),  [`Polybar #2810`](https://github.com/polybar/polybar/pull/2810))  by [@madhavpcm](https://github.com/madhavpcm).
+  - Added `format-hover` and `label-hover` to the base module template so any module can easily make use of this with `m_formatter->add_default_formats()`
 - Network unit type ([Polybar PR #3001 by @Cationiz3r](https://github.com/polybar/polybar/pull/3001))
   - `internal/network`: `metric-units` setting
 - reverse-scroll, logarithmic-scroll and click-toggle ([Polybar PR #2703 by me](https://github.com/polybar/polybar/pull/2703))